# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: GroupBuy Comments API
  version: 1.0.0
tags:
  - name: Comments
paths:
  /api/comments:
    get:
      tags:
        - Comments
      operationId: listComments
      summary: List comments
      parameters:
        - name: purchaseItemId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Comments
      operationId: createComment
      summary: Create comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/comments/{id}:
    get:
      tags:
        - Comments
      operationId: getComment
      summary: Get comment by id
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '200':
          description: Comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Comments
      operationId: updateComment
      summary: Update comment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Comments
      operationId: deleteComment
      summary: Delete comment
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '204':
          description: Comment deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    CommentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CommentDto:
      type: object
      required:
        - id
        - purchaseItemId
        - authorId
        - authorDisplayName
        - content
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        purchaseItemId:
          type: integer
          format: int64
        authorId:
          type: integer
          format: int64
        authorDisplayName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      required:
        - purchaseItemId
        - content
      properties:
        purchaseItemId:
          type: integer
          format: int64
        content:
          type: string
          minLength: 1
          maxLength: 2000
    UpdateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
    ErrorResponse:
      type: object
      properties:
        message:
          type: string